collection_Arraylist_studentnaming_percentage
The Collection in Java is a framework that provides an architecture to store and manipulate the group of objects. Java Collections can achieve all the operations that you perform on a data such as searching, sorting, insertion, manipulation, and deletion. Java Collection means a single unit of objects. An Iterator is an object that can be used to loop through collections, like ArrayList and  hashlist. It is called an "iterator" because "iterating" is the technical term for looping. The ArrayList class is a Java class that you can use to store lists of objects. You can also store objects in an array, but arrays have a couple of obvious problems. To create an array, you have to specify a size for the array.